Split lorebooks panel helpers
